Consolidate VMStructs and JVMFlag classes
